[WIP]: Use a sprase file backing
The downside to this implementation is that it requires four syscalls
(mmap, mmap, munmap, munmap) for every comparison. This implementation
may be useful for other cases where the overhead of syscalls is not the
bottleneck. It may also be possible to refine this implementation to
reduce the number of syscalls but no effort to do this has been made.